# internal/api/openapi/swagger.yaml
basePath: /api
definitions:
  api.APIError:
    properties:
      code:
        type: integer
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
  api.StatusResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    REST API for the RepLog workout tracking application.

    Authentication uses session cookies issued by `POST /api/login`
    (or the WebAuthn / token-login endpoints). Cookies are
    `HttpOnly`, `SameSite=Lax`, and (in production) `Secure`.
    The browser sends them automatically on subsequent requests.
  title: RepLog API
  version: 1.0.0
paths:
...
  /auth/token/{token}:
    get:
      description: Single-use token issued by an admin (POST /users/{userID}/tokens).
        Creates a session on success.
      parameters:
      - description: Login token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User + redirect URL
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Log in via magic-link token
      tags:
      - Auth
...
  /setup/passkey/skip:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Skip passkey setup
      tags:
      - Auth
swagger: "2.0"
...